Rename commuted_calligraphy package to telegraphy
