+++ pyproject.toml
@@ -15,13 +15,13 @@ dependencies = ["PyYAML>=6.0.3"]
 dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.8.0"]
 
 [project.scripts]
-story-brief = "commuted_calligraphy.story_brief.generate_story_brief:main"
+story-brief = "telegraphy.story_brief.generate_story_brief:main"
 
 [tool.setuptools.packages.find]
-include = ["commuted_calligraphy*"]
+include = ["telegraphy*"]
 
 [tool.setuptools.package-data]
-"commuted_calligraphy.story_brief.data" = ["*.json"]
+"telegraphy.story_brief.data" = ["*.json"]
 
 [tool.ruff]
 line-length = 100
