+++ telegraphy/story_brief/data/entities.json
