+++ telegraphy/story_brief/data/prompts.json
