+++ tests/conftest.py
@@ -13,7 +13,7 @@ from typing import Any
 
 import pytest
 
-from commuted_calligraphy.story_brief import generate_story_brief as story_brief
+from telegraphy.story_brief import generate_story_brief as story_brief
 
 _STORY_DATASET_FILES = tuple(story_brief.STORY_DATASET_FILES.values())
