+++ tests/story_brief/test_availability_filters.py
@@ -2,7 +2,7 @@ from datetime import date
 
 import pytest
 
-from commuted_calligraphy.story_brief import generate_story_brief as story_brief
+from telegraphy.story_brief import generate_story_brief as story_brief
 
 
 def test_available_characters_filters_by_date(monkeypatch: pytest.MonkeyPatch) -> None:
