+++ tox.ini
@@ -10,9 +10,9 @@ setenv =
     COVERAGE_PROCESS_START = {toxinidir}/tox.ini
     COVERAGE_FILE = {toxworkdir}/.coverage
     TOX_PROJECT_ROOT = {toxinidir}
-    PYTHONPATH = {toxinidir}{pathsep}{toxinidir}/commuted_calligraphy/scripts/cov_init
+    PYTHONPATH = {toxinidir}{pathsep}{toxinidir}/telegraphy/scripts/cov_init
 commands =
-    python -m commuted_calligraphy.scripts.run_coverage_workflow
+    python -m telegraphy.scripts.run_coverage_workflow
 
 [testenv:py312-fast]
 deps =
@@ -27,7 +27,7 @@ commands =
     python -m pytest -m "slow or integration" tests
 
 [coverage:run]
-source = commuted_calligraphy
+source = telegraphy
 relative_files = True
 parallel = True
 branch = True
